Added launchfiles for each stage (launch sim, spawn robot)
