+++ andino_webots/launch/spawn_andino_webots.launch.py
@@ -1,0 +1,201 @@
+# BSD 3-Clause License
+
+# Copyright (c) 2023, Ekumen Inc.
+# All rights reserved.
+
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+
+# 1. Redistributions of source code must retain the above copyright notice, this
+#    list of conditions and the following disclaimer.
+
+# 2. Redistributions in binary form must reproduce the above copyright notice,
+#    this list of conditions and the following disclaimer in the documentation
+#    and/or other materials provided with the distribution.
+
+# 3. Neither the name of the copyright holder nor the names of its
+#    contributors may be used to endorse or promote products derived from
+#    this software without specific prior written permission.
+
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+# FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+# DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+# SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+# OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+import os
+import pathlib
+import launch
+import xacro
+
+from launch import LaunchDescription
+from launch.actions import DeclareLaunchArgument
+from launch.conditions import IfCondition
+from launch_ros.actions import Node
+from launch.substitutions import LaunchConfiguration
+from ament_index_python.packages import get_package_share_directory
+from webots_ros2_driver.urdf_spawner import URDFSpawner, get_webots_driver_node
+from webots_ros2_driver.webots_controller import WebotsController
+from webots_ros2_driver.wait_for_controller_connection import (
+    WaitForControllerConnection,
+)
+
+
+def configure_gazebo_sensors(robot_description: str):
+    """
+    Configures the lidar's near parameter and attaches the camera to a rotated frame so that
+    """
+    robot_description = robot_description.replace(
+        "</ray>",
+        """
+                    <clip>
+                        <near>0.05</near>
+                    </clip>
+                </ray>
+        """,
+    )
+    robot_description = robot_description.replace(
+        '<gazebo reference="camera_link">', '<gazebo reference="webots_camera_link">'
+    )
+    return robot_description
+
+
+andino_webots_pkg_dir = get_package_share_directory("andino_webots")
+andino_control_pkg_dir = get_package_share_directory("andino_control")
+
+
+def generate_launch_description():
+    use_sim_time = LaunchConfiguration("use_sim_time")
+
+    use_rsp = DeclareLaunchArgument(
+        "rsp",
+        default_value="true",
+        description="Select whether to spawn the Robot State Publisher",
+    )
+    use_sim_time_arg = DeclareLaunchArgument(
+        "use_sim_time",
+        default_value="true",
+    )
+
+    andino_gazebo_xacro_path = os.path.join(
+        andino_webots_pkg_dir, "urdf", "andino_webots_description.urdf.xacro"
+    )
+    andino_gazebo_description = xacro.process_file(
+        andino_gazebo_xacro_path,
+        mappings={"use_gazebo_ros_control": "False", "use_fixed_caster": "False"},
+    ).toprettyxml(indent="    ")
+    andino_gazebo_description = configure_gazebo_sensors(andino_gazebo_description)
+
+    # TODO(#12): Update to PROTOSpawner when implementation is released
+    spawn_andino = URDFSpawner(
+        name="andino",
+        robot_description=andino_gazebo_description,
+        translation="0 0 0.022",
+        rotation=" 0 0 1 0",
+    )
+
+    # Robot state publisher
+    params = {"robot_description": andino_gazebo_description, "publish_frequency": 30.0}
+    rsp = Node(
+        package="robot_state_publisher",
+        executable="robot_state_publisher",
+        namespace="",
+        output="both",
+        parameters=[params],
+        condition=IfCondition(LaunchConfiguration("rsp")),
+    )
+
+    andino_robot_driver = Node(
+        package="webots_ros2_driver",
+        executable="driver",
+        output="screen",
+        additional_env={"WEBOTS_CONTROLLER_URL": "Andino Webots"},
+        parameters=[
+            {"robot_description": andino_gazebo_description},
+        ],
+    )
+    # Webots Controller to initialize cameras/LIDARs
+    andino_webots_path = os.path.join(
+        andino_webots_pkg_dir, "urdf", "andino_webots.urdf"
+    )
+    ros2_control_params = os.path.join(
+        andino_control_pkg_dir, "config", "andino_controllers.yaml"
+    )
+    mappings = [
+        ("/diff_controller/cmd_vel_unstamped", "/cmd_vel"),
+        ("/diff_controller/odom", "/odom"),
+    ]
+
+    andino_webots_controller = WebotsController(
+        robot_name="andino",
+        parameters=[
+            {
+                "robot_description": andino_webots_path,
+                "use_sim_time": use_sim_time,
+            },
+            ros2_control_params,
+        ],
+        remappings=mappings,
+        respawn=True,
+    )
+
+    # ROS2 control
+    controller_manager_timeout = ["--controller-manager-timeout", "5000"]
+    diffdrive_controller_spawner = Node(
+        package="controller_manager",
+        executable="spawner",
+        output="screen",
+        arguments=["diff_controller"] + controller_manager_timeout,
+        parameters=[
+            {"use_sim_time": use_sim_time},
+        ],
+    )
+    joint_state_broadcaster_spawner = Node(
+        package="controller_manager",
+        executable="spawner",
+        output="screen",
+        arguments=["joint_state_broadcaster"] + controller_manager_timeout,
+    )
+    ros_control_spawners = [
+        diffdrive_controller_spawner,
+        joint_state_broadcaster_spawner,
+    ]
+
+    # Wait for the simulation to be ready to start the diff drive and spawners
+    waiting_nodes = WaitForControllerConnection(
+        target_driver=andino_webots_controller, nodes_to_start=ros_control_spawners
+    )
+
+    return LaunchDescription(
+        [
+            use_sim_time_arg,
+            # Request to spawn the Andino via URDF
+            spawn_andino,
+            # Add andino's controller
+            andino_webots_controller,
+            waiting_nodes,
+            # Robot state publisher
+            use_rsp,
+            rsp,
+            # Launch the driver node once the URDF robot is spawned
+            launch.actions.RegisterEventHandler(
+                event_handler=launch.event_handlers.OnProcessIO(
+                    target_action=spawn_andino,
+                    on_stdout=lambda event: get_webots_driver_node(
+                        event, andino_robot_driver
+                    ),
+                )
+            ),
+            # Kill all the nodes when the driver node is shut down (useful with other ROS 2 nodes)
+            launch.actions.RegisterEventHandler(
+                event_handler=launch.event_handlers.OnProcessExit(
+                    target_action=andino_robot_driver,
+                    on_exit=[launch.actions.EmitEvent(event=launch.events.Shutdown())],
+                )
+            ),
+        ]
+    )
